feat: rename package to nvim-artisan
