+++ lua/nvim-artisan/init.lua
@@ -1,11 +1,11 @@
 local M = {}
 
-local builtins = require("artisan.utils.builtins")
-local plugin_loader = require("artisan.plugin-loader")
+local builtins = require("nvim-artisan.utils.builtins")
+local plugin_loader = require("nvim-artisan.plugin-loader")
 
 function M:run()
     -- Set options
-    require("artisan.options")
+    require("nvim-artisan.options")
 
     -- Disable builtin providers and plugins
     builtins.disable_plugins()
@@ -15,10 +15,10 @@ function M:run()
     plugin_loader.init()
 
     -- Load plugins
-    plugin_loader.load(require("artisan.plugins"))
+    plugin_loader.load(require("nvim-artisan.plugins"))
 
     -- Set mappings
-    require("artisan.mappings")
+    require("nvim-artisan.mappings")
 end
 
 return M
